+++ data-raw/DATASET.R
@@ -69,6 +69,21 @@ futuresMonths <- dplyr::tibble(Month = seq.Date(as.Date("2022-01-01"),as.Date("2
                                Code = c("F","G","H","J","K","M","N","Q","U","V","X","Z"))
 usethis::use_data(futuresMonths, overwrite = T)
 
+# Futures specifications
+library(rvest)
+rD <- rsDriver(browser = "firefox", port = 4545L, verbose = F)
+remDr <- rD[["client"]]
+Sys.sleep(2)
+remDr$navigate("https://www.cmegroup.com/trading/energy/crude-oil/light-sweet-crude_contract_specifications.html")
+Sys.sleep(2)
+#remDr$findElement(using = 'class', value = 'productTemplate')$clickElement()
+remDr$findElement(using = 'class', value = 'section')$clickElement()
+page <- remDr$getPageSource()
+futuresSpecs <- read_html(page[[1]]) %>% rvest::html_table(fill = TRUE) %>% .[[1]] %>%
+  dplyr::select(Specification = X1, Description = X2)
+remDr$close()
+usethis::use_data(futuresSpecs, overwrite = T)
+
 ## steo for RTLappWTI
 
 steo <- RTL::chart_eia_steo(key = EIAkey)